framework WLANRTSYS-11255: build on Linux support
Elaborate:
Add support for building all multiap repos on Linux, currently with dummy bpl. Not tested on Linux.
Building on Linux requires ubus (and therefore lua & ubox) which can be built and installed manually on a linux machine.
Eventually if bpl will indeed be required, it can be implemented with something other than ubus, but then changes will be needed
in the other repos.

Changed:
- Added dummy bpl for linux, currently ~empty, since it is used by agent/common/controller
- Updated README.md with ubus requirement

Test:
- Build on Linux
- Cross-compile for RDKB and UGW
- Runtime test on UGW

Dependencies: Yes (common/controller/agent)
Signed-off-by: Tomer Eliyahu <tomer.b.eliyahu@intel.com>