optimal_path_task::work: remove switch fallthroughs
The state machine switch in optimal_path_task::work() has two
fallthrough cases. Both of them are in fact OK as fallthrough, because
the following case label is indeed the state that we go to. However, it
makes more sense to yield on every state change. So add a break.

Detected with -Werror=implicit-fallthrough

Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>