beerocks_cli: make readline dependency optional
Since readline is GPL-3.0+ licensed, we don't want to force a dependency
on it. If we did, any binary distribution of beerocks_cli would
effectively be under GPL-3.0+.

Make readline optional by using getline() instead of readline() if
readline is not enabled. There is no support for history and completion
in this case, of course.

Everything that is specific to readline is put under ifdef
HAVE_READLINE.

readline() and getline() are wrapped in a new function
getline_interactive() that also takes care of freeing the buffer after
converting it to a string.

Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>