maptools: build: fix parallel build on older CMake
The -j option for cmake --build was only introduced in CMake 3.12. To
support older CMake versions, replace it with a -j option passed after
the --.

Unfortunately, ninja doesn't support -j without argument. Therefore,
explicitly specify the number of jobs. One more than the number of CPUs
sounds like a good starting point.

Note that this limits the build engines to Ninja and Make, because
others don't know the -j option at all.

Similarly, the -v option for cmake --build was only introduced in CMake
3.14. There is no easy solution for that one, so keep using the original
VERBOSE=1 there. This breaks the ninja build, so verbose has to be done
with make.

Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>