cmake: restructure CMakeLists.txt to make it a single project
This will allow to build everything in the regular CMake way. Follow-up
commits will refine things for a better out-of-the box experience.

All cmake files are collected in the top-level cmake directory, which is
the one that gets included in the search path.

Create a top-level CMakeLists.txt that contains the common parts. It
adds the four components as subdirectories. All the pieces of the
top-level CMakeLists.txt files that are used in at least two components
are moved here.

Add a bunch of TODO entries for future improvements.

Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>