build: fix cmake RPATH settings
We normally build and install the framework to a none standard path. In
this case, when running installed binaries, the linker needs to know
where the shared libraries used by the binary are located.

There are several ways the linker searches for libraries -
LD_LIBRARY_PATH, /etc/ld.so.conf, builtin directories (/lib, /usr/lib,
etc.).

RPATH is a list of directories which are searched first and are "linked"
to the executable, which allow setting the search paths during build.

CMAKE provides full control over the RPATH for executables and
libraries, but it needs to be told to set RPATH since by default no
RPATH is used.

This commit fixes RPATH settings globally in the top level
CMakeLists.txt to always produce RPATH for installed binaries/libraries
unless the installation folder is a system folder (in which case it is
not needed).

For more reading - https://gitlab.kitware.com/cmake/community/wikis/doc/cmake/RPATH-handling

Dependencies - None
Tests -
 - Build (full build with MSGLIB=nng BUILD_TESTS=ON)
 - Run tests without the need for setting LD_LIBRARY_PATH
 - Verify correct nng used by built binaries (LD_TRACE_LOADED_OBJECTS=1)

Signed-off-by: Tomer Eliyahu <tomer.b.eliyahu@intel.com>