build: fix PythonTlvf use
Since tlvf.py is now used internally, it is not yet installed so it
can't be found by find_program. Instead, make the PythonTlvf variable
CACHED INTERNAL (which makes it global) so it can also be used from the
common module.

Remove the check for PythonTlvf being set in common.

Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>