easylogging: add S_IROTH to elpp file permissions
Add S_IROTH to file permissions so that created folders and log files
can be read by other users. Original permissions have write permissions
for other, so the only reason not to give read permissions to others is
to prevent log files read by attackers which may contain sensitive data.
But even in this case the better solution is to simply _not_ store
sensitive data in log files.

Signed-off-by: Tomer Eliyahu <tomer.b.eliyahu@intel.com>