Remove all find_package calls
Now everything is internal to the project, there is no longer a need to
have find_package calls to find the (installed) libraries.

elpp is kept as an "external" package, even though it is built as part
of the normal process. This will allow to move it out at some point.