Fix include paths of internal libraries
Some of the internal headers used to be installed in a different
location than their source location, e.g. with a beerocks/bwl prefix.
Since we no longer export the libraries, they can no longer be included
from such a path.

Therefore, the #include statement must be stripped of such prefixes.

In addition, the proper include location must be defined in the PUBLIC
part of the INCLUDE_DIRECTORIES property.