tools: move build folder outside of prplMesh repo
Before merging prplMesh repos into one repository, the build folder was
created in the top directory, where .repo folder exists.
This was done in order to prevent accidental removal when using git
clean -xdf, and to prevent having untracked folders.

Now that we have a single repository (prplMesh), the build folder is
created in prplMesh repo - so update build.py to create the folder in
the parent directory.

Signed-off-by: Tomer Eliyahu <tomer.b.eliyahu@intel.com>