tlvf: fixing old yamls of ApAutoConfig flow
tlvApRadioIdentifier - change incorrect struct type used for mac address

tlvApRadioBasicCapabilities - this tlv uses nested lists and was written
(but not used) before the support in nested list was added to tlvf.
Therefore, the structure of this tlv was wrong. this change fixes its
structure to be compatible with nested lists standard as tlvf requires.

Signed-off-by: Lior Amram <lior.amram@intel.com>