framework/tlvf: re-instate tlvf test
The tlvf test was disabled in CMake, and also no longer built.
Re-instate it now.

Remove the debug messages in the beginning (the beerocks part didn't
build, hlen is anyway not used, and the other message is not very
useful).

NOTE: this test runs into the same issue as the ieee1905_transport test:
it doesn't terminate if using nng as the msglib.

Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>