framework/tlvf/test: remove local bus
The local bus wasn't testing anything, since a message was published on
it but nobody reads it. Also, testing the local bus is adequately
covered by the ieee1905_transport_test.

Remove the local bus from the tlvf test.

This also avoids the issue that the test doesn't terminate when nng is
used as broker.

Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>