framework/tlvf/test: remove ieee1905 CmduTxMessage
The CmduTxMessage wasn't testing anything; it was only used as a buffer,
but then the buffer was just copied to the receive buffer for parsing.
Also, testing CmduTxMessage is adequately covered by the
ieee1905_transport_test.

Remove the CmduTxMessage from the tlvf test.

This removes the dependency on ieee1905_transport, so the test can be
built without any msglib.

Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>