maptools.py: build safelibc with larger strings
safelibc puts an artificial limit on the maximum string size. This limit
is set at configuration time and defaults to 4K. However, we need to
handle much larger strings - dwpal's HOSTAPD_TO_DWPAL_MSG_LENGTH is set
to 16K.

For a bit of safety margin, set the maximum string size to 64K.

In order to do this, the acbuilder class is extended with an additional
field extra_config_args with extra arguments to be passed to the
configure script.

Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>