ieee1905.1: TLV 'type' field is 1 byte only
IEEE 1905.1 defines the TLV as a 1-byte type, 2-byte length and
variable-length value. However, the TLV definition in the prplMesh
source code used 2 bytes (uint16_t) for the type.

With the 'type' field at 1 byte, the default struct packing is incorrect
(the compiler will insert an extra byte to make sure the length field is
16-bit aligned). Therefore, we also must make sure the structs are
packed correctly, by adding the 'packed' attribute (cfr. #18).

Also, the TLV structure layout is repeated in a few places, so all of
these have to be fixed:

1. Socket thread implements cmdu_verify which does a simple sanity on the
CMDU. It also suffered from the same mistake of declaring the type as 16
bit, so fix that.

2. tlvf base class (non-autogenerated) - TLV header size fix from 4
bytes to 3 bytes (since type field is 1 byte) in cBaseClass CmduMessage
classes. This solves addClass failing on 'not enough buffer space'.

3. Fix eTlvfMap to be 1 byte instead of 2

4. Remove 16bit swap for tlv type field

Finally, update all autogenerated files with the 'type' field fix.

Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>
Signed-off-by: Tomer Eliyahu <tomer.b.eliyahu@intel.com>