prplMesh-platform-init.sh: remove radio from network-manager
If network-manager is running on the host, it will manipulate the radios
as soon as they appear. This can sometimes block hostapd from starting.

To avoid this, tell network-manager not to consider the mac80211_hwsim
devices. Look up those devices in sysfs.

While we're at it, also set the number of radios explicitly when the
module is probed.

Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>