tools: add scripts for running in container
Add scripts to run agent and controller in containers.

make-container copies the required shared libraries and a few additional
binaries from the host system into the build/container directory. That
directory will be the root for the container.

start-container starts the container using systemd-nspawn. The script
also contains the hack needed to work around systemd-nspawn's limitation
that it doesn't support wireless network interfaces. start-container
will launch a script within the container.

The containers are ephemeral, i.e. nothing is persistent. This is needed
anyway, otherwise it's not possible to run two containers from the same
directory.

To make sure we do get some output from the containers, the log files
are put in a directory build/log-${script} which gets bind-mounted into
the container.

Two scripts are provided: container-run-agent and
container-run-controller.

Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>