CMakeLists: add -Wno-address-of-packed-member CFLAG
Recent GCC (at least GCC 9, maybe already earlier) adds a warning for
taking a possibly unaligned address of a member of a packed struct.
Since we use packed structs all over the place now, we hit this warning
in a few places.

However, all the platforms we target (x86, ARM, MIPS) support unaligned
accesses. So this warning is not really important.

Therefore, suppress it with the -Wno-address-of-packed-member flag. Even
though only the agent and common components really need it, apply it
globally.

Note that older GCC versions don't know this flag. However, GCC silently
ignores and -Wno- flag it doesn't know about.

Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>