tlvf: remove AutoGenerated files
The AutoGenerated files are currently included in the git repository.
This has the disadvantage that it obfuscates any change to the tlvf yaml
files or the copied source files, because the changes in AutoGenerated
tend to be much larger.

Also, there's a risk that the commited AutoGenerated files are not
consistent with the committed yaml files. This has happened already.

There are two reasons to include the AutoGenerated files in the git
repository.

* It allows to build on a platform that doesn't have Python or can't run
  tlvf.py for some other reason. However, that scenario is not even
  supported by our CMakeLists.txt.

* It makes it possible for people to browse the complete code online.

The second reason is the only valid one, and it is not considered strong
enough compared to the disadvantages. It is expected that people who
want to browse the code will checkout the repository and also build it.
This is consistent with how many other projects with generated source
files do it.

Therefore, remove the AutoGenerated directories in framework/tlvf and
common/beerocks/tlvf.

Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>