tlvf: CmduMessage::getNextTlvType(): TLV type is 8 bit
Commit 4136c80f8ee76 fixed the IEEE1905 TLV 'type' field from 16 bits to
8 bits. However, parsing the TLV 'type' field is also open-coded in
CmduMessage::getNextTlvType(). It reads 16 bits from the buffer and
swaps them.

Similarly, CmduMessage::getNextTlvLength() takes an offset of 2 bytes
into the buffer, which should be only 1 byte.

The tlvf_test breaks because of this. This was not detected since it is
not yet executed in CI.

Fix both CmduMessage::getNextTlvType() and
CmduMessage::getNextTlvLength() to use only one byte for the TLV `type'
field.

Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>