.travis.yml: check that AutoGenerated is up to date
It is possible (and it has happened before) that the AutoGenerated files
are not up-to-date in a pull request.

Let Travis check for this, simply by checking the git status after the
build. If the AutoGenerated files are outdated, there will be changes in
git. `git diff-index HEAD` is the plumbing command to check for such
changes. If nothing changed, it prints nothing; if there are changes, it
prints machine-readable information about modified files and returns
non-zero.

Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>