tools: set GID as well as UID
Commit 5e83aec7ec made sure that the build is done as the current user
instead of as root. However, it did not make sure that the group is also
set to the user's group. Thus, the AutoGenerated files and the files in
the build/ directory will belong to group 'root'.

Similar to what we do for uid, add a variable gid and set it based on
SUDO_GID, and use it in the USER directive.

Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>