common: scripts: install per target
Currently we install scripts which are only used by cross platforms
(RDKB, UGW) such as beerocks_utils.sh, beerocks_watchdog.sh, etc.

On the Linux platform, only install prplmesh_utils.sh. The other scripts are
still installed for UGW and RDKB. Although the intention is to also replace these
with prplmesh_utils.sh, we'll do that once that script is ready for it.

Signed-off-by: Tomer Eliyahu <tomer.b.eliyahu@intel.com>