github: add issue templates
The issue templates feature by github allows us to define templates from
which people can select when creating an issue. It is also one of the
missing bullets we have in the "community" tab in github, which says
that we are "missing something".

Add templates for bug reports, feature requests and tasks.

Signed-off-by: Tomer Eliyahu <tomer.b.eliyahu@intel.com>
Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>