framework: set zmq the default messaging library
Since currently nng cannot be used, and since the controller
certification milestone does not need to go on a real product, and since
zmq is linked to but is expected to be in the system (not built by
prplMesh), use zmq as the default messaging library.
In a real product, zmq will not be available and cmake variable MSGLIB
will need to be overriden.

While we're at it - make MSGLIB cached variable with one of the
following string options - "zmq", "nng", "None" so it can be selected in
`cmake-gui`.

Signed-off-by: Tomer Eliyahu <tomer.b.eliyahu@intel.com>