tools: docker builder build native os option
In some cases we want to build using the same OS version as in the host.
For example, consider the case that the host is ubuntu-19.04, where some
of the packages which are installed are newer than the default image
from which the builder derives (ubuntu-18.04).
This means that we won't necessarilly be able to run the images built by
the builder from the host.

This is a major issue though since we can also run from docker using the
runner docker, but adding an option to build the base image using the
same OS version is nice, so this commit adds it.

See issue #72

Signed-off-by: tbeliyah <tomer.b.eliyahu@intel.com>