tools: docker runner: option to use host os
Add an option to build an image based on the host OS version instead of
the hard coded Ubuntu-18.04.

This is needed in some cases when we want to be able to run prplMesh
built on the host, so the shared libraries (for example json-c) can
differ between ubuntu versions.

This commit adds an optional flag to the image-build.sh script to use
the host OS version instead of always using ubuntu-18.04.

Signed-off-by: tbeliyah <tomer.b.eliyahu@intel.com>