tools: docker runner: remove installdir from image
Having the install directory in the image may cause accidently
using the wrong directory as WORKDIR in case the image was
built on a different prplMesh clone.

Remove installdir from the image, move it to the entry point.
In addition, add environment variable INSTALL_DIR to point on the actual
install dir and use it in the helper scripts start-agent, start-
controller and start-controller-agent.

Signed-off-by: tbeliyah <tomer.b.eliyahu@intel.com>