tools: docker automatic image build fix
runner and builder docker helper scripts support building the images
automatically if are not found. However, since the last change, the
path for ./image-build.sh has changed, so if images are not prebuilt the
scripts fail.
This commit fixes this and updates the README.md by moving the scripts
one directory up to `prplMesh/tools/docker`.

Signed-off-by: tbeliyah <tomer.b.eliyahu@intel.com>