controller: autoconfig M1 - use radio mac as key
The radio mac of the radio agent is available in M1 message, so use it
as the main key for the controller DB node representing this radio.
This allows using the handle_slave_join() as is, which contains the RUID
which later will also be taken from the WSC TLV.
The same node in the DB can be reached by using the source mac (AL MAC)
of a CMDU + the RUID: <al_mac>_<ruid> (not a new addition).

Signed-off-by: Tomer Eliyahu <tomer.b.eliyahu@intel.com>