scripts: prplmesh_utils.sh platform_db support
Add platform_db automatic creation by prplmesh_utils.sh according to the
mode in which the script is invoked, and write that to
/tmp/prplmesh_platform_db.

While we're at it, change eth0 to sim-eth0 since if
prplmesh_utils.sh runs on the host, it will mess up networking.
Using a ~unique name for the dummy interface is good practice anyhow,
and allows us to run the same script in both containers and hosts.
Changing wlan0/2 to sim-wlan0/2 will be done in a separate commit.

Signed-off-by: Tomer Eliyahu <tomer.b.eliyahu@intel.com>