common: refactor net_struct sMacAddr
sMacAddr was used for MAC Address AND for scan results. This is just wrong,
so change sMacAddr to be just a MAC address
struct, and add a new struct for the scan results.
While we're at it - add operator== and operator!= implementations
to all of the structs defined in net_struct.h.

Changes caused changes in the common, agent and controller,
and are all part of this commit.

Signed-off-by: Tomer Eliyahu <tomer.b.eliyahu@intel.com>