framework: common: add encryption - diffie_hellman
Create a new source and include for encryption in the common library.

Create a class diffie_hellman to do diffie_hellman exchange, as well as
the corresponding test program. The test program is built even if MSGLIB
is None - as opposed to the other tests.

Add OpenSSL to the build.

Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>