CI: update .travis.yaml for operational test
Update .travis.yaml to run operational test during continuous
integration - running 2 nodes, each in a separate container - controller
and agent node, and an agent only node, connected through a docker
bridge.
Testing is using prplmesh_utils.sh status command which checks that all
processes are running, and that both the main agent and radio agents are
operational.
For the agent only node, there is still an issue so we don't fail the
test on this.

Note that we build in debug mode since release build is broken in
containers - see issue
https://github.com/prplfoundation/prplMesh/issues/93

Signed-off-by: Tomer Eliyahu <tomer.b.eliyahu@intel.com>