coding style: Update clang-format.sh
Update clang-format.sh to use clang-format available in the system.
Till now we forced using clang-format-3.9 since this is what we have in
Intel build servers, but since anyway we need to run with root
permissions we are no longer using these servers.
Also, running with the default clang-format will allow CI to also run
clang-format and block PRs which did not.

Signed-off-by: Tomer Eliyahu <tomer.b.eliyahu@intel.com>