framework: tlvf: add getMessageId() API
When parsing a CMDU, the header is never swapped (CmduMessageRx::parse)
since it may be parsed multiple times when forwarded over UDS.
For that reason, any CMDU header field needs to have a special getter
function that will handle endianess.
Currently, there is a single such getter -
`CmduMessage::getMessageType()`.

This commit adds `CmduMessage::getMessageId()` which is needed when the
upper layer needs to reply with a CMDU with the same MID.

Signed-off-by: Tomer Eliyahu <tomer.b.eliyahu@intel.com>