common: btl: set preset_message_id before bus send
When the upper layer needs to send a CMDU with the same MID as in a
received CMDU, it will set the MID in the CMDU header, and the transport
in this case will not generate a new MID but keep the same one in the
CMDU header.

But this works only if the upper layer marks this in the
metadata()->preset_message_id - otherwise the transport will override
the MID.

This commit sets the preset_message_id to be true if there is a none
zero MID in the CMDU header. Since this is a CmduMessageTx, the only one
who can set the MID is the sender.

Signed-off-by: Tomer Eliyahu <tomer.b.eliyahu@intel.com>