framework: tlvf: refactor to use single sMacAddr
There are 2 structs in the codebase describing a MAC address - one in
the BCL, beerocks::net::sMacAddr, and the other in the TLV factory,
sMacAddress. Refactor the code to use a single struct, the one defined
in the TLV factory.

Changed the struct name to match shorter name set in BCL, and change the
bytes array name to oct to prevent conflicts. After that simply remove
the namespace prefix and we're done.

No point in splitting this commit to auto generated and yaml since each
commit must not break, and this will break it.

Main files which were changed:
- bcl net_struct.h - removed beerocks::net::sMacAddr and move compare
operators outside the namespace
- tlvf sMacAddress.yaml - renamed to sMacAddr.yaml, changed the
struct name to sMacAddr and changed the byte array from mac[] to oct[].
This is for reducing the conflicts in all of the code base to almost
zero after removing the namespace prefixes.

Signed-off-by: Tomer Eliyahu <tomer.b.eliyahu@intel.com>