tools: docker: add multinode test gw+repeater
Add a wrapper script which starts a gateway node (controller+agent) in
one container, a repeater node (agent) in another container, waits a bit
then tests that both are operational.

This script can be easilly used by developers (and in the future CI) to
quickly see if anything got broken.

Signed-off-by: Tomer Eliyahu <tomer.b.eliyahu@intel.com>