tlvf: add getNextTlvType() API
Add API in CmduMessage class to get the next TLV type.
This is needed in the following cases:
- Unknown TLVs - for example an unknown VS TLV that should simply be
skipped
- Optional TLVs - self explanatory, some TLVs are optional so we should
first check if they are available in the message
- Looped back messages - Some Message types are used for both agents and
controller, and due to the local bus architecture, a message sent from
the controller for instance will also be received by the controller
since it is subscribed for the same message type. An example for such
messages are the AP autoconfiguration WSC message - message sent from
the agent have one AP Radio Basic capabilities TLV followed by one WSC
M1 TLV. a message sent from the controller contains one AP Radio
Identifier TLV followed by one or more WSC M2 TLVs. In these cases we
can add a simple call to getNextTlvType and check if this is a looped
back message.

Signed-off-by: Tomer Eliyahu <tomer.b.eliyahu@intel.com>