backhaul mgr: remove extra connected notification
Once the backhaul manager successfully connects to the controller after
the first received autoconfig search response, it sends a notification
to the agent which then starts the master join flow (autoconfig WSC).
The agent in question is the first agent which joined the backhaul
manager and is marked as the "backhaul manager slave".

The backhaul manager function which sends the notification is
finalize_slaves_connect_state(), which sends the
cACTION_BACKHAUL_CONNECTED_NOTIFICATION notification to all
registered slaves regardless if the notification has already been sent
to that slave or not. The reason for that is for cases where the
backhaul manager restarted or got disconnected from the controller, in
which case all slaves should rejoin the controller (AKA start autoconfig
WSC again).

It relies on the fact that slaves support backhaul disconnection in
which case the slaves should re-join the master, and mark it as
'reconfiguration' in the vendor specific data (see `is_backhual_reconf`
in the son_slave_thread.cpp).

The outcome of this is that there will always be 2 backhaul connected
notificaitons to the backhaul manager slave, which results with 2
autoconfig WSC M1 messages, one of which is redundant.

In order to prevent that, add a member to the backhaul_manager_thread
class per slave socket - operational_on_registration, which is set when
a radio agent (slave) registers to the backhaul manager after it is
already connected or operational. Then, only for this case (backhaul
reconnect), resend the connected notification to the agent.

This solves the issue where 2 M1 messages are sent from the same slave,
which might fail us in future agent certification and just looks odd and
raises questions when looking at current wireshark capture.

Signed-off-by: Tomer Eliyahu <tomer.b.eliyahu@intel.com>