tlvf: add getBuffRemainingBytes() overload API
Add getBuffRemainingBytes(void *start) API which will be used in
followup commits for adding variable length lists support in TLVF.
The new API simply returns the number of bytes from the start pointer
till the buffer end. This will be used to `memmove` data on allocation
of variable length list.

Signed-off-by: Tomer Eliyahu <tomer.b.eliyahu@intel.com>