tlvf: fix missing swap for external enums
When a  class contains an enum member which is defined in an external
yaml file, it's type is not identified as ENUM by the tlvf.py
generateObject() method, which is incorrect and causing the script
not to generate swapping logic for this member.
In case where the enum is not a single byte, it is a real bug.

Fix this by adding logic in generateObject() to correctly identify
ENUM members even if they are in an external namespace.

Signed-off-by: Tomer Eliyahu <tomer.b.eliyahu@intel.com>