tlvf: variable length list support
There are 2 types of variable (dynamic) length lists in the TLV factory:
- known length of the format `list[ LIST_SIZE ]`
- unknown length of the format `list[ ]`

From the programmer point of view, these are almost identical. The main
difference is that known length lists have the length variable part of
the TLV.
An Unknown length list has to be the last member in the TLV so its length
can be calculated using the TLV length field.

Variable length lists with known length (`list [ LIST_SIZE ]`) also
needs to be the last member of the TLV class due to insufficient support
in the TLV factory for variable length lists.
The reason for that is because all of the TLV class members are written
directly on the buffer (the class itself holds pointers to each member
on the buffer), so allocating a new list element for a list which is not
the last requires an update to all of the following members pointers,
and moving all data to the new "END" of the list.

This limitation is currently preventing us from properly implementing
some types of TLVs, such as M1 and M2 WSC which have variable length
strings such as 'manufacturer', 'device_name', etc.

This commit adds full support for variable length lists as follows:
Variable length members have an `alloc` method which is used to allocate
room on the buffer for the newly created list member.
On every such `alloc` function, tlvf.py will now also add a `memmove`
call to move all the remainder bytes following the current list "END" to
the new list "END". The number of bytes to move is simply calculated
using `getBuffRemainingBytes(void *start)` new API which was added for
this in a previous commit.
In addition, we also adds a "marker" (special comment used by the TLV
factory to mark locations to which we add new code when adding following
class members) to mark the location in the `alloc` function to which to
add code for each following member to update its pointer.

In case the list elements are of class type, for example:
```
cOuter:
  list_length:
    _type: uint16_t
    _length_var: True
  list:
    _type: cInner
    _length: [ list_length ]

cInner:
  list_length:
    _type: uint16_t
    _length_var: True
  list:
    _type: cInner
    _length: [ list_length ]
```

We also update the create method with a similar `memmove` since the
create method adds the class on the buffer with the initial size, before
the allocation method is called.

Signed-off-by: Tomer Eliyahu <tomer.b.eliyahu@intel.com>