tlvf: align M1 and M2 to WSC specification
The EasyMesh Multi-AP specification defines the autoconfig flow which
uses WiFi Simple Setup (WSC) specification v2.0.6 M1 and M2 messages for
agent and controller fAPs and bAPs configuration.

The WSC specification defines the size of each parameter in the M1 and
M2 TLVs. Some of the attributes have fixed size, while others, the
stging attributes - 'manufacturer', 'model_name', 'device_name`,
`serial_number` and `model_number` - all are defined as variable length
lists of type CHAR as follows:
```
manufacturer_length:
  _type: uint16_t
  _length_var: True
manufacturer:
  _type: char
  _length: [ manufacturer_length ]
```

Due to the previous lack of support in the TLV factory for variable
length lists, these were defined with fixed length, which is not aligned
with the specification and will cause prplMesh to fail in certification
if the string sizes are not the masimum allowed by the WSC spec as we
defined them in the `WSC_attributes.yaml` file.
The outcome would be, in a very high probability, that prplMesh will not
pass certification and will not be able to work with other vendors which
implemented this correctly.

Since the previous commits added variable length lists support, this
commit updates the M1 and M2 TLVs to use variable length lists instead
of fixed sized strings, thus aligning to the WSC specification.

In addition, this commit also fixes the following:
- aligns the vendor extension attribute structure to the Multi-AP spec
- Move vendor extension definitions to TLVF defined in eWscVendorExt.yaml
- Removes sM1 and sM2 yamls since now we are using tlvWscM1 and tlvWscM2

Note - TLVF still lacks support for defining a variable length list if
the list member type is a class without an internal variable length
list, therefore all the variable length strings defined in the WSC
specification are moved to the TLVs themeselves.
This actually makes the usage cleaner, so not a high priority issue.
Issue is tracked in
https://github.com/prplfoundation/prplMesh/issues/115

Signed-off-by: Tomer Eliyahu <tomer.b.eliyahu@intel.com>