agent: fix order of list creatin in M1
Due to the previous commit which added protection against adding lists
out of order, the agent WSC M1 message failed to add the serial number
attribute in the WSC M1 TLV since it was done out of order.
This commit changes the order of variables creation and fixes the issue.

Actually this protection is not really needed in this case since only if
the list member type is another complex list, a buffer corruption may
occur.
But adapting the TLVF python script to support full out of order
assignments is too complex at this stage of the project.
The alternative of removing the protection is also bad since it is far
better to have a warning print where its not really necessary and
failing the list allocation than having corrupt buffer.

While at it - also add validity checks for each assignment and return
false on error, in both son_master_thread and son_slave_thread.

Signed-off-by: Tomer Eliyahu <tomer.b.eliyahu@intel.com>