build: unify CMAKE find_package() calls
Move all common find_package() calls for the controller, agent and
common which finds the installed framework and common packages to
multiap-helpers.cmake in each module (controller, agent and common).

This is done now since both the controller and agent now needs to link
to mapf::common since they will need the encryption support which will
be added there.

Signed-off-by: Tomer Eliyahu <tomer.b.eliyahu@intel.com>