tlvf: fix swap length in init when not needed
Commit
https://github.com/prplfoundation/prplMesh/pull/114/commits/c5f4548b33f63dbfcf4621382c3d401a2a8651d0
added a swap for variable length list length member.
This was needed since on Init, when parsing a TLV, the byte order of the
length field is the network byte order and the swap function is only
called at the end of the init function.
However, we need the real value in order to set the correct pointers
which come after the list.

However 2, which is the reason for this commit, is that on some
occasions, swap is not needed since the message is already swapped - for
example messages which are sent over UDS, therefore m_swap__ is set to
false.
This commit adds a condition to only swap the value if m_swap__ and
m_parse__ are True, same condition which later is used before init
function completes.

Signed-off-by: Tomer Eliyahu <tomer.b.eliyahu@intel.com>