tlvf: Fix type checking for none uint8_t types
Although most types are 1 byte, some can be
larger - for example, the encrypted settings in the
M2 TLV is 2 bytes long.
The current validity check in the TLVF generated
init() function is done before endianity conversion is executed,
which is find if the type is 1 byte, but causes the
init() function to fail if the type is larger.
Move the type checking to after the swap call
to prevent that.

Signed-off-by: Tomer Eliyahu <tomer.b.eliyahu@intel.com>