controller: bml: Drill down WFA-CA command to the controller
1. Remove unnecessary function that parses WFA-CA command and printed
to the console dummy return string.

2. Drill down WFA-CA string command from bml to the controller.

3. Added command with response callback mechanism.
According to the Test Suit Control API specifications, UCC should get
two strings in return. The first on is "Running" which means that the
system received the command, and it started to do what has been requested,
and the second return is a "Complete" status that means that the system
succeded to perform what it was requested.
All of the replies which will be received to the will be routed to callback
function, which will print it.

4. Add new wfa_ca class to concentrate all WFA-CA activity.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>