documentation: update README.md to install clang-format
clang-format is a utility which does the formatting of the prplMesh code,
using ./clang-format.sh at the root of the repository.
Using clang-format is mandatory for all developers - and CI enforces that by running
clang-format on each PR and each commit inside a PR, and fail if changes are there.
Therefore, add clang-format to README.md install.

Signed-off-by: pavel levitsky <pavellev@inter.net.il>