framework: tlvf: move _is_tlv_class to obj_meta
If _is_tlv_class=True, the class with the same name as the file is
treated as a TLV class, meaning its type is checked, and the length
member is updated as part of the TLVF operation whenever the length of
the class changes.

In order to support definitions of TLV classes in different file names
then the class name, move the metadata defining a TLV class into the
object meta instead of declaring it at the top of the YAML.

Update all TLVs yaml files in the TLVF accordingly.

Using this will allow adding the encrypted settings attribute to the
WSC_Attributes.yaml instead of declaring it in a separate file.

Signed-off-by: Tomer Eliyahu <tomer.b.eliyahu@intel.com>