tlvf: fix swapping of simple list of scalars
The swapping function for the entries of a list of scalars was missing
an address-of.

This issue was not detected before because the only simple lists that
are actually used have uint8_t members, which don't need swapping. All
other lists have class members which have a swap function so they don't
need an ampersand.

Also, the compiler doesn't detect this because we use reinterpret_cast.

Fix this by adding the address-of. We also add it for classes, even
though it's never actually added there because classes have swap_is_func
== True.

Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>