tlvf: tlvfTestVarList: use 1 byte for list lengths
The tlvf currently doesn't support a list length of more than one byte,
because it handles swapping incorrectly. Since we don't actually need
list lengths with more than one byte, it will not be fixed now.
Therefore, just change tlvfTestVarList do always use one byte for the
list length.

The problem lies in the class_swap function: it will first swap the
length member and then use this length member to iterate over the
elements. For parsing, this is indeed correct. For sending, however,
after swapping the length would no longer be correct, so the length has
to be swapped after iterating over the list instead of before. For
1-byte length this is not a problem since it doesn't need to be swapped.

Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>